requirements.txt
absl-py==2.3.1
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
Bootstrap-Flask==2.2.0
Brotli==1.1.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
cmake==4.0.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.15
decorator==5.2.1
deep-translator==1.11.4
defusedxml==0.7.1
Django==5.2
django-crispy-forms==2.4
django-extensions==4.1
django-jazzmin==3.0.1
django-logentry-admin==1.1.0
dlib @ file:///C:/dlib/Dlib_Windows_Python3.x-main/Dlib_Windows_Python3.x-main/dlib-19.24.99-cp312-cp312-win_amd64.whl#sha256=20c62e606ca4c9961305f7be3d03990380d3e6c17f8d27798996e97a73271862
docopt==0.6.2
elevenlabs==2.22.0
executing==2.2.0
face-recognition==1.3.0
face_recognition_models @ git+https://github.com/ageitgey/face_recognition_models@e67de717267507d1e9246de95692eb8be736ab61
fastapi==0.115.13
fastjsonschema==2.21.1
Flask==2.3.2
Flask-CKEditor==0.4.6
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==25.2.10
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.7.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.50.0
google-generativeai==0.8.5
googletrans==4.0.0rc1
greenlet==3.1.1
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==5.2.0
idna==2.10
ipykernel==6.30.0
ipython==8.12.3
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.7.1
jaxlib==0.7.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.7
langchain-text-splitters==1.0.0
langsmith==0.4.44
lark==1.2.2
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mediapipe==0.10.21
mistune==3.1.3
ml_dtypes==0.5.3
multidict==6.5.0
mypy_extensions==1.1.0
mysql-connector==2.2.9
mysql-connector-python==9.4.0
narwhals==1.37.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
ngrok==1.5.1
nltk==3.9.1
notebook==7.4.4
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.3.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
orjson==3.11.4
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.2.1
pipreqs==0.5.0
platformdirs==4.3.8
playsound==1.2.2
plotly==6.0.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycares==4.9.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.12.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==6.3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-json-logger==3.3.0
pytz==2025.1
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.30.0
Send2Trash==1.8.3
sentencepiece==0.2.1
sentry-sdk==2.29.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soupsieve==2.6
SpeechRecognition==3.14.3
spotipy==2.25.1
SQLAlchemy==2.0.25
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.46.2
tenacity==9.1.2
terminado==0.18.1
textblob==0.19.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
twilio==9.8.5
types-python-dateutil==2.9.0.20250708
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.34.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.0
wsproto==1.2.0
WTForms==3.0.1
yarg==0.1.9
yarl==1.20.1
zstandard==0.25.0
